refactor: 👷folder /componentProps to /props
