+++ components/layout/page-navigation/PageNavigation.tsx
@@ -7,7 +7,7 @@ import { DATA_TEST_IDS } from '@/__tests__/playwright/lib/utils/constants/ids/da
 
 import { NavigationDirectionEnum } from '@/lib/utils/typeDefinitions/enums'
 
-import { PageNavigationProps } from '@/lib/utils/typeDefinitions/componentProps'
+import { PageNavigationProps } from '@/lib/utils/typeDefinitions/props'
 
 const PageNavigation: FC<PageNavigationProps> = ({
   linkPrevious,
