+++ components/pages/projects/other-experience/OtherExperienceContent.tsx
@@ -5,7 +5,7 @@ import Heading4 from '@/components/shared/Heading4'
 
 import { TEXT } from '@/localization/english'
 
-import { OtherExperienceContentProps } from '@/lib/utils/typeDefinitions/componentProps'
+import { OtherExperienceContentProps } from '@/lib/utils/typeDefinitions/props'
 
 const OtherExperienceContent: FC<OtherExperienceContentProps> = ({ experience }): JSX.Element => {
   return (
