+++ components/pages/projects/overview-page/ProjectSection.tsx
@@ -3,7 +3,7 @@ import { FC } from 'react'
 import ExperienceSection from '@/components/pages/projects/other-experience/ExperienceSection'
 import ProjectSummaryCard from '@/components/pages/projects/overview-page/ProjectSummaryCard'
 
-import { ProjectSectionProps } from '@/lib/utils/typeDefinitions/componentProps'
+import { ProjectSectionProps } from '@/lib/utils/typeDefinitions/props'
 
 const ProjectSection: FC<ProjectSectionProps> = ({ projectData, sectionId, sectionText }): JSX.Element => {
   return (
