+++ components/pages/testimonials/TestimonialDescription.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 
 import Paragraph from '@/components/shared/Paragraph'
 
-import { TestimonialDescriptionProps } from '@/lib/utils/typeDefinitions/componentProps'
+import { TestimonialDescriptionProps } from '@/lib/utils/typeDefinitions/props'
 
 const TestimonialDescription: FC<TestimonialDescriptionProps> = ({ description }) => (
   <Paragraph>{description}</Paragraph>
