+++ components/shared/Breadcrumbs.tsx
@@ -9,7 +9,7 @@ import { DATA_TEST_IDS } from '@/__tests__/playwright/lib/utils/constants/ids/da
 import { ARIA_LABELS } from '@/lib/utils/constants/ariaLabels'
 import { PAGES_URL } from '@/lib/utils/constants/urls/pageUrls'
 
-import { BreadCrumbsProps } from '@/lib/utils/typeDefinitions/componentProps'
+import { BreadCrumbsProps } from '@/lib/utils/typeDefinitions/props'
 
 const BreadCrumbs: FC<BreadCrumbsProps> = ({ level1Url, level1Text, level2Url, level2Text }): JSX.Element => {
   {
