+++ components/shared/DividerWithText.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-import { DividerWithTextProps } from '@/lib/utils/typeDefinitions/componentProps'
+import { DividerWithTextProps } from '@/lib/utils/typeDefinitions/props'
 
 const DividerWithText: FC<DividerWithTextProps> = ({ text }): JSX.Element => {
   return (
