+++ components/shared/Heading5.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 
 import { getDataTestIdAttribute } from '@/lib/utils/helpers/getDataTestIdAttribute'
 
-import { Heading5Props } from '@/lib/utils/typeDefinitions/componentProps'
+import { Heading5Props } from '@/lib/utils/typeDefinitions/props'
 
 const Heading5: FC<Heading5Props> = ({ children, dataTestId = '', textColor = '', customCss = '' }): JSX.Element => {
   return (
