+++ components/shared/Highlight.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react'
 
-import { HighlightProps } from '@/lib/utils/typeDefinitions/componentProps'
+import { HighlightProps } from '@/lib/utils/typeDefinitions/props'
 
 const Highlight: FC<HighlightProps> = ({ text }): JSX.Element => {
   return <span className="font-bold text-violet-600">{text}</span>
